Memory usage for temporary arraies are logged.
git-svn-id: https://svn.sourceforge.jp/svnroot/molds/MolDS/trunk@410 1136aad2-a195-0410-b898-f5ea1d11b9d8

+++ src/zindo/ZindoS.h
@@ -124,7 +124,7 @@ private:
                                         double** residualVector, 
                                         double** ritzVector) const;
    void FreeDavidsonRoopCISTemporaryMtrices(double*** interactionMatrix, 
-                                            double interactionMatrixDimension, 
+                                            int interactionMatrixDimension, 
                                             double** interactionEigenEnergies) const;
 };
